+++ Disassemble8080.h
@@ -3,5 +3,5 @@
 #include <string>
 
 namespace Disassemble8080 {
-	size_t dissassembleOpcode(uint8_t* code, std::string& outOpcode);
+	uint16_t dissassembleOpcode(uint8_t* code, std::string& outOpcode);
 }
